evl: kconfig: introduce high per-CPU concurrency switch
EVL_HIGH_PERCPU_CONCURRENCY optimizes the implementation for
applications with many real-time threads running concurrently on any
given CPU core (typically when eight or more threads may be sharing a
single CPU core). This is a combination of the scalable scheduler and
rb-tree timer indexing as a single configuration switch, since both
aspects are normally coupled.

If the application system runs only a few EVL threads per CPU core,
then this option should be turned off, in order to minimize the cache
footprint of the queuing operations performed by the scheduler and
timer subsystems. Otherwise, it should be turned on in order to have
constant-time queuing operations for a large number of runnable
threads and outstanding timers.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>